Fixed Bugs and added some new ones
Actually what I did, fixed some bugs that only showed with less specific tests in bishop, Queen, and pawn. Added a more complete move class with state and if a piece was taken, and castling. I also did a final checkMove in Game, which checks for a legal move and if the move would put them in check.  However, adding the Move to moveList causes errors for some reason, with an array out of bounds, for some reason that is beyond my comprehension